+++ libddog/dashboards/__init__.py
@@ -38,7 +38,7 @@ from libddog.dashboards.enums import (
     TitleAlign,
     VerticalAlign,
 )
-from libddog.dashboards.layouts import HLayout, HLayoutStack
+from libddog.dashboards.layouts import HLayout, HLayoutStack, HLayoutWrapping
 from libddog.dashboards.presets import NotePreset
 from libddog.dashboards.widgets import (
     Group,
@@ -61,6 +61,7 @@ __all__ = (
     "Group",
     "HLayout",
     "HLayoutStack",
+    "HLayoutWrapping",
     "LayoutType",
     "LegendColumn",
     "LegendLayout",